requirements.txt
# Python 3.10+ required
flask>=3.0.0
flask-cors>=4.0.0
groq>=0.4.0
edge-tts>=6.1.0
python-dotenv>=1.0.0
